pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tinyfilter"
version = "0.1.6"
authors = [
  { name="Zephaniah Roe", email="zroe@uchicago.edu" },
]
description = "tinyfilter is the computer vision equivalent to micrograd. It convert images into ASCII art using the principles of CNNs (convolutional neural networks)."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "numpy ~= 1.25.0",
  "Pillow",
]

[project.urls]
"Homepage" = "https://github.com/zroe1/tinyfilter/tree/main"

[project.scripts]
tinyfilter = "tinyfilter:tiny_print"
